Break apart view.ts into multiple files.
